DZ5/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Сервис авторизации"
authors = ["Ekaterina <batova.ekateriina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.39"
pydentic = "^0.0.1.dev3"
uvicorn = "^0.34.0"
python-jose = "^3.4.0"
passlib = "^1.7.4"
pydantic-settings = "^2.8.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"
psycopg2 = "^2.9.10"
redis = {extras = ["asyncio"], version = "^4.6.0"}



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
